msm_fb: display: workaround a single underrun when lcdc is on
Develop a workaround for a single underrun instance which has no
obvious s/w reasons. The justification behind this workaround is that
when this single underrun happens, all pipes are intended to be solid
fill mode because framebuffer is not ready yet to be displayed. So,
the thing that really matters is to make sure this underrun dose not
cause any visual side effects by identifying when this underrun
happens and matching the underrun color to pipe solid color. And after
that the underrun color can be restored.

Change-Id: I039625c0522306492c4cefd7ca5ec1d14e7e9c0a
Signed-off-by: Huaibin Yang <huaibiny@codeaurora.org>